Was not a good idea to move the DB connection because includes it to unit tests => fix it
