+++ src/main/app.js
@@ -1,21 +1,9 @@
 const express = require('express');
-const mongoose = require('mongoose');
 
 const zikResourceRouter = require('./zikresource/zikresource-api');
 
 const app = express();
 
-// TODO : change to use env var
-const connect = mongoose.connect('mongodb://localhost:27017/zikstock',
-    {
-        useNewUrlParser: true,
-        useUnifiedTopology: true,
-        useCreateIndex: true,
-        useFindAndModify: false
-    });
-connect.then(() => {
-    console.log("Connected to MongoDB");
-}, (err) => { console.log(err); });
 
 app.use('/api/zikresources', zikResourceRouter);
